# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/features": {
            "post": {
                "description": "Создает новую фичу на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать новую фияу",
                "operationId": "create-feature",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен администратора",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для создания фичи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/features.RequestFeature"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданная фича",
                        "schema": {
                            "$ref": "#/definitions/features.ResponseFeature"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании фича",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        },
...
        "/tags": {
            "post": {
                "description": "Создает новый тег на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать новый тег",
                "operationId": "create-tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен администратора",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для создания тега",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tags.RequestTag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный тег",
                        "schema": {
                            "$ref": "#/definitions/tags.ResponseTag"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании тега",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        },
        "/user_banner": {
            "get": {
                "description": "Получает баннер с учетом переданных параметров запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить баннер с указанными параметрами",
                "operationId": "get-banner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен пользователя",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор особенности баннера",
                        "name": "feature_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор тега баннера",
                        "name": "tag_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Использовать последнюю ревизию",
                        "name": "use_last_revision",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Содержимое баннера",
                        "schema": {
                            "$ref": "#/definitions/banners.Content"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "404": {
                        "description": "Баннер не найден",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Создает нового пользователя на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать нового пользователя",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "Данные для создания пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.RequestUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseUser"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании пользователя",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "banners.Content": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "banners.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "features.RequestFeature": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "features.ResponseFeature": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "feature_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "tags.RequestTag": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "tags.ResponseTag": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tag_id": {
                    "type": "integer"
                }
            }
        },
        "users.RequestUser": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "users.ResponseUser": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}